+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  * Add `librenms_location` resource support
  * Update `librenms_devicegroup.devices` to Set instead of List
  * Add device `display` and `location` attributes
+ * Fix snmp_ attr handling in device resource
  * Refresh go-librenms to fix some float64 handling
  * Add test scenarios